tried to return model (doesn't work)

+++ sandbox_bench/bench_lib/src/lib.rs
@@ -22,16 +22,8 @@ extern crate once_cell;
 
 // static mut GLOBAL: u64 = 0;
 
-// #[AlohomoraSandbox()]
-// pub fn global_test(input: String) -> u64 {
-//   unsafe {
-//     let x: u64 = GLOBAL;
-//     GLOBAL += 1;
-//     x
-// }
-// }
-
 #[derive(Debug, Sandboxable)]
+// #[repr(Rust)]
 #[repr(C)]
 pub struct Test {
   pub a: i32,
@@ -44,72 +36,11 @@ pub struct Test {
   // pub ptr2: *const Test2,
 }
 
-// #[cfg(not(target_arch = "wasm32"))]
-// pub struct TestUnswizzled {
-//   pub a: <i32 as Sandboxable>::InSandboxUnswizzled, 
-//   pub b: <usize as Sandboxable>::InSandboxUnswizzled,
-//   pub c: <isize as Sandboxable>::InSandboxUnswizzled,
-//   pub t: <Test2 as Sandboxable>::InSandboxUnswizzled,
-//   pub ptr: <*mut Test2 as Sandboxable>::InSandboxUnswizzled,
-// }
-
-// #[derive(Sandboxable)]
-// pub struct TestR {
-//   pub a: i32,
-//   pub b: usize,
-//   pub c: isize,
-//   pub t: Test2,
-//   pub ptr: *mut Test2,
-// }
-
 #[derive(Clone, Debug, Sandboxable)]
 #[repr(C)]
 pub struct Test2 {
   pub a: usize,
 }
-
-// #[cfg(not(target_arch = "wasm32"))]
-// pub struct Test2Unswizzled {
-//   pub a: u32,
-// }
-
-// #[cfg(not(target_arch = "wasm32"))]
-// impl<T> Sandboxable for Test<T> {
-//   type InSandboxUnswizzled = TestUnswizzled;
-//   fn into_sandbox(outside: Self, alloc: alohomora_sandbox::alloc::SandboxAllocator) -> Self::InSandboxUnswizzled {
-//       TestUnswizzled {
-//         a: Sandboxable::into_sandbox(outside.a, alloc.clone()),
-//         b: Sandboxable::into_sandbox(outside.b, alloc.clone()),
-//         c: Sandboxable::into_sandbox(outside.c, alloc.clone()),
-//         t: Sandboxable::into_sandbox(outside.t, alloc.clone()),
-//         ptr: Sandboxable::into_sandbox(outside.ptr, alloc.clone()),
-//       }
-//   }
-//   fn out_of_sandbox(inside: &Self::InSandboxUnswizzled, any_sandbox_ptr: usize) -> Self {
-//       Test {
-//         a: Sandboxable::out_of_sandbox(&inside.a, any_sandbox_ptr),
-//         b: Sandboxable::out_of_sandbox(&inside.b, any_sandbox_ptr),
-//         c: Sandboxable::out_of_sandbox(&inside.c, any_sandbox_ptr),
-//         t: Sandboxable::out_of_sandbox(&inside.t, any_sandbox_ptr),
-//         ptr: Sandboxable::out_of_sandbox(&inside.ptr, any_sandbox_ptr),
-//       }
-//   }
-// }
-
-// #[cfg(not(target_arch = "wasm32"))]
-// impl Sandboxable for Test2 {
-//   type InSandboxUnswizzled = Test2Unswizzled;
-//   fn into_sandbox(outside: Self, alloc: alohomora_sandbox::alloc::SandboxAllocator) -> Self::InSandboxUnswizzled {
-//     Test2Unswizzled{
-//       a: Sandboxable::into_sandbox(outside.a, alloc),
-//     }
-//   }
-//   fn out_of_sandbox(inside: &Self::InSandboxUnswizzled, any_sandbox_ptr: usize) -> Self {
-//       Test2 {
-//         a: Sandboxable::out_of_sandbox(&inside.a, any_sandbox_ptr),
-//       }
-//   }
-// }
 
 // Sandbox functions.
 #[AlohomoraSandbox()]
@@ -168,7 +99,10 @@ pub fn train2(inputs: Vec<(NaiveDateTime, u64)>) -> (u64, (), u64) {
 
 #[AlohomoraSandbox()]
 // pub fn train(inputs: (Vec<(NaiveDateTime, u64)>, u64)) -> (u64, FittedLinearRegression<f64>, u64) {
-pub fn train(inputs: Vec<(NaiveDateTime, u64)>) -> (usize, (), usize) {
+pub fn train(inputs: Vec<(NaiveDateTime, u64)>) -> (u64, (), u64) {
+  // TODO: can't return FittedLinearRegression<f64> here bc 
+  // its actually very complicated and hard to derive `Sandboxable` for
+  
   // END TIMER (start in bin)
   // let start = Utc::now().timestamp_nanos_opt().unwrap() as u64;
   // println!("in the sandbox, inputs are {:?}", inputs.len());
